# Integration of HyDE for Enhanced Document Retrieval in the RAG Pipeline by Group 19