--- a/server/.env.template
+++ b/server/.env.template
@@ -96,3 +96,8 @@ ragas_answer_instruction="You are a digital librarian and need to answer questio
 ragas_answer_query="Answer the following question, never give any explanation or other output than the generated article itself:
 
 {question}"
+
+hyde_enabled=True
+hyde_multi_generations=1
+hyde_prompt_template="Robustness in LLMs"
+domain="Robustness in LLMs"