--- a/server/.env.template
+++ b/server/.env.template
@@ -99,5 +99,6 @@ ragas_answer_query="Answer the following question, never give any explanation or
 
 hyde_enabled=True
 hyde_multi_generations=1
-hyde_prompt_template="Robustness in LLMs"
+hyde_task_type="Scientific inquiry"
 domain="Robustness in LLMs"
+hyde_context_instruction="Provide a concise scientific explanation for the following query in the context of {context}."