--- a/server/.env.template
+++ b/server/.env.template
@@ -101,4 +101,4 @@ hyde_enabled=True
 hyde_multi_generations=1
 hyde_task_type="Scientific inquiry"
 domain="Robustness in LLMs"
-hyde_context_instruction="Provide a concise scientific explanation for the following query in the context of {context}."
+hyde_context_instruction="Write a scientific paper passage to answer the question in the context of {context}.\n Your response must be direct and avoid any unnecessary phrases, personal remarks, or repetitive text."