--- a/server/.env.template
+++ b/server/.env.template
@@ -55,7 +55,7 @@ rag_question_followup="The follow-up question you have to answer:
 rag_fetch_new_instruction="Instruction: You are a digital librarian with a database that contains relevant documents for user queries. Users want to ask questions based on those documents and ask questions that either need you to fetch new documents from the database or that are a followup question on previously obtained documents. You need to decide whether you are going to fetch new documents or whether the user is asking a follow-up question but you don't get to see the actual documents the user potentially is looking at.\nShould new documents be fetched from the database based on this user query? Answer with yes or no."
 rag_fetch_new_question="The user question is the following: \"{question}\"\n"
 
-use_rewrite_loop=True
+use_rewrite_loop=False
 rewrite_query_instruction="You have to answer a user question based on documents retrieved from a document database. It is your task to decide whether or not the documents contain the answer to the user's query. You can always only answer with exactly yes or no. The documents that are currently fetched from the database are:
 
 {context}"
@@ -97,8 +97,19 @@ ragas_answer_query="Answer the following question, never give any explanation or
 
 {question}"
 
+# Enable or disable HyDE functionality
 hyde_enabled=True
+
+# Number of generations for HyDE
 hyde_multi_generations=1
-hyde_task_type="Scientific inquiry"
-domain="Robustness in LLMs"
-hyde_context_instruction="Write a scientific paper passage to answer the question in the context of {context}.\n Your response must be direct and avoid any unnecessary phrases, personal remarks, or repetitive text."
+
+# Generalized HyDE template for flexible and reusable prompts
+hyde_general_template="Write a {context_type} passage to {action}. Include {additional_context}.\nYour response must be direct and avoid any unnecessary phrases, personal remarks, or repetitive text.\n"
+
+# Custom HyDE template for specific use cases (leave empty to use general template)
+hyde_custom_template=""
+
+# Default values to populate the generalized HyDE template
+hyde_default_context_type="scientific"
+hyde_default_action="answer the following question"
+hyde_default_additional_context="Please provide references and ensure an academic tone"