@ErikTromp 2024-12-05
Could you remove the data folder?